requirements.txt
python-telegram-bot>=20.0
APScheduler
openai-whisper
python-dotenv
psutil
requests
aiohttp
aiosqlite
ffmpeg-python
torch
numpy
fastapi
uvicorn
jinja2
pyautogui
pyperclip
beautifulsoup4
cryptography
pyzipper
feedparser
pyshorteners
textblob
geopy
websockets>=12.0
